intl: Rename package to "translations"
