+++ src/main/kotlin/com/emberjs/translations/EmberIntl.kt
@@ -1,4 +1,4 @@
-package com.emberjs.intl
+package com.emberjs.translations
 
 import com.emberjs.Ember
 import com.intellij.openapi.vfs.VirtualFile
