+++ src/main/kotlin/com/emberjs/translations/JsonStringPropertyCollector.kt
@@ -1,4 +1,4 @@
-package com.emberjs.intl
+package com.emberjs.translations
 
 import com.intellij.json.psi.JsonProperty
 import com.intellij.json.psi.JsonStringLiteral
