+++ src/main/resources/META-INF/plugin.xml
@@ -56,7 +56,7 @@
 
         <testFinder implementation="com.emberjs.navigation.EmberTestFinder"/>
 
-        <fileBasedIndex implementation="com.emberjs.intl.EmberIntlIndex"/>
+        <fileBasedIndex implementation="com.emberjs.translations.EmberIntlIndex"/>
 
         <psi.referenceContributor language="JavaScript" implementation="com.emberjs.psi.EmberReferenceContributor"/>
