+++ src/test/kotlin/com/emberjs/translations/EmberI18nTest.kt
@@ -1,26 +1,26 @@
-package com.emberjs.intl
+package com.emberjs.translations
 
 import com.intellij.testFramework.fixtures.LightPlatformCodeInsightFixtureTestCase
 import com.intellij.util.indexing.FileBasedIndex
 import org.assertj.core.api.Assertions.assertThat
 import java.nio.file.Paths
 
-class EmberIntlTest : LightPlatformCodeInsightFixtureTestCase() {
+class EmberI18nTest : LightPlatformCodeInsightFixtureTestCase() {
 
     override fun getTestDataPath(): String? {
-        val resource = ClassLoader.getSystemResource("com/emberjs/intl/fixtures")
+        val resource = ClassLoader.getSystemResource("com/emberjs/translations/fixtures")
         return Paths.get(resource.toURI()).toAbsolutePath().toString()
     }
 
-    fun testFindBaseLocale() {
+    fun testFindDefaultLocale() {
         // Load fixture files into the project
-        myFixture.copyDirectoryToProject("ember-intl-with-base-locale", "/")
+        myFixture.copyDirectoryToProject("ember-i18n", "/")
 
         // Rebuild index now that the `package.json` file is copied over
         FileBasedIndex.getInstance().requestRebuild(EmberIntlIndex.NAME)
 
-        val psiFile = myFixture.configureByFile("app/templates/base-locale-test.hbs")
+        val psiFile = myFixture.configureByFile("app/templates/application.hbs")
 
-        assertThat(EmberIntl.findBaseLocale(psiFile)).isEqualTo("de")
+        assertThat(EmberI18n.findDefaultLocale(psiFile)).isEqualTo("zh")
     }
 }
