+++ src/test/resources/com/emberjs/translations/fixtures/ember-i18n/app/locales/en/translations.js
