+++ src/test/resources/com/emberjs/translations/fixtures/ember-i18n/app/locales/zh/translations.js
