+++ src/test/resources/com/emberjs/translations/fixtures/ember-i18n/app/templates/application.folded.hbs
