+++ src/test/resources/com/emberjs/translations/fixtures/ember-i18n/config/environment.js
