+++ src/test/resources/com/emberjs/translations/fixtures/ember-intl-json/app/templates/json-expectation.hbs
