+++ src/test/resources/com/emberjs/translations/fixtures/ember-intl-json/package.json
