+++ src/test/resources/com/emberjs/translations/fixtures/ember-intl-json/translations/en.json
