+++ src/test/resources/com/emberjs/translations/fixtures/ember-intl-with-base-locale/app/templates/base-locale-test-expectation.hbs
