+++ src/test/resources/com/emberjs/translations/fixtures/ember-intl-with-base-locale/translations/de.yaml
