+++ src/test/resources/com/emberjs/translations/fixtures/ember-intl/app/templates/folding-test.hbs
