+++ src/test/resources/com/emberjs/translations/fixtures/ember-intl/app/templates/missing-translation-folding-test-expectation.hbs
