+++ src/test/resources/com/emberjs/translations/fixtures/ember-intl/app/templates/placeholder-folding-test-expectation.hbs
