+++ src/test/resources/com/emberjs/translations/fixtures/ember-intl/app/templates/sexpr-folding-test-expectation.hbs
