+++ src/test/resources/com/emberjs/translations/fixtures/ember-intl/package.json
