+++ src/test/resources/com/emberjs/translations/fixtures/ember-intl/translations/de.yaml
