+++ src/test/resources/com/emberjs/translations/fixtures/no-dependencies/app/templates/folding-test.hbs
