+++ src/test/resources/com/emberjs/translations/fixtures/no-dependencies/package.json
