+++ src/test/resources/com/emberjs/translations/fixtures/no-dependencies/translations/en.yaml
